requirements.txt
pycairo
svg.path==3.0

pylint
